makes a second layer of function passing.
example:
dispatch(astNode, function (val, ret) {
  return ret(val)
})

When traversing it allows a node transformer to select which part should
be affected by its parent. The ret function is called on the part that
would normally be returned.  Often this means return ret(...).  By
passing this in instead of doing it on return the child nodes have more
control over how the parent wraps them.  This is good if the child
transforms in a way that the parent wouldn't expect. I think this
simplifies the logic?